+++ FormScore.cs
@@ -1,13 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
 using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
 using System.Data.SqlClient;
+using System.Windows.Forms;
 
 namespace QLSV
 {
@@ -51,12 +45,12 @@ namespace QLSV
 
         private void btn_add_Click(object sender, EventArgs e)
         {
-            connect.Open();           
+            connect.Open();
             SqlCommand cmdCheck = new SqlCommand(@"SELECT MASV FROM SINHVIEN WHERE MASV = @id", connect);
             cmdCheck.Parameters.AddWithValue("@id", txt_id.Text);
-            SqlDataReader dr = cmdCheck.ExecuteReader();      
+            SqlDataReader dr = cmdCheck.ExecuteReader();
             if (dr.Read())
-            { 
+            {
                 connect.Close();
                 connect.Open();
                 SqlCommand com = new SqlCommand(@"INSERT INTO DIEMSV (MASV,MAMH,DIEM) VALUES (@id,@sub_id,@score)", connect);
@@ -68,7 +62,7 @@ namespace QLSV
 
 
             }
-            else               
+            else
             {
                 if (string.IsNullOrEmpty(txt_id.Text) == true || string.IsNullOrEmpty(txt_sub_id.Text) == true || string.IsNullOrEmpty(txt_score.Text) == true)
                 {
@@ -80,24 +74,38 @@ namespace QLSV
                     MessageBox.Show("Mã sinh viên không tồn tại");
                     connect.Close();
                 }
-                
+
             }
             connect.Close();
             loadData();
+
+            txt_id.ResetText();
+            txt_sub_id.ResetText();
+            txt_score.ResetText();
         }
         private void btn_del_Click(object sender, EventArgs e)
         {
             connect.Open();
-            SqlCommand com = new SqlCommand(@"DELETE FROM DIEMSV WHERE MASV = @id AND MAMH=@sub_id ", connect);
-            com.Parameters.AddWithValue("@id", txt_id.Text);
-            com.Parameters.AddWithValue("@sub_id", txt_sub_id.Text);
-            com.ExecuteNonQuery();
-            MessageBox.Show("Xóa thành công");
+            if (string.IsNullOrEmpty(txt_id.Text) == true || string.IsNullOrEmpty(txt_sub_id.Text) == true || string.IsNullOrEmpty(txt_score.Text) == true)
+            {
+                MessageBox.Show("Phải chọn sinh viên cần xóa");
+                connect.Close();
+            }
+            else
+            {
+                SqlCommand com = new SqlCommand(@"DELETE FROM DIEMSV WHERE MASV = @id AND MAMH=@sub_id ", connect);
+                com.Parameters.AddWithValue("@id", txt_id.Text);
+                com.Parameters.AddWithValue("@sub_id", txt_sub_id.Text);
+                com.ExecuteNonQuery();
+                MessageBox.Show("Xóa thành công");
+                connect.Close();
+            }
+            
             loadData();
-            txt_id.Text = "";
-            txt_sub_id.Text = "";
-            txt_score.Text = "";
-            connect.Close();
+            txt_id.ResetText();
+            txt_sub_id.ResetText();
+            txt_score.ResetText();
+
         }
 
         private void btn_update_Click(object sender, EventArgs e)
@@ -105,26 +113,37 @@ namespace QLSV
             connect.Open();
             txt_id.ReadOnly = true;
             txt_sub_id.ReadOnly = true;
-            SqlCommand com = new SqlCommand(@"UPDATE  DIEMSV SET DIEM=@score WHERE MASV =@id AND MAMH=@sub_id", connect);
-            com.Parameters.AddWithValue("@id", txt_id.Text);
-            com.Parameters.AddWithValue("@sub_id", txt_sub_id.Text);
-            com.Parameters.AddWithValue("@score", double.Parse(txt_score.Text));
-            com.ExecuteNonQuery();
-            MessageBox.Show("Cập nhật thành công");
+            if (string.IsNullOrEmpty(txt_id.Text) == true || string.IsNullOrEmpty(txt_sub_id.Text) == true || string.IsNullOrEmpty(txt_score.Text) == true)
+            {
+                MessageBox.Show("Phải chọn sinh viên cần cập nhật");
+                connect.Close();
+            }
+            else
+            {
+                SqlCommand com = new SqlCommand(@"UPDATE  DIEMSV SET DIEM=@score WHERE MASV =@id AND MAMH=@sub_id", connect);
+                com.Parameters.AddWithValue("@id", txt_id.Text);
+                com.Parameters.AddWithValue("@sub_id", txt_sub_id.Text);
+                com.Parameters.AddWithValue("@score", double.Parse(txt_score.Text));
+                com.ExecuteNonQuery();
+                MessageBox.Show("Cập nhật thành công");
+                connect.Close();
+            }
+            
             loadData();
-
-            connect.Close();
+            txt_id.ResetText();
+            txt_sub_id.ResetText();
+            txt_score.ResetText();
         }
 
         private void btn_reset_Click(object sender, EventArgs e)
         {
-            txt_id.ReadOnly = false;
-            txt_id.Text = "";
-            txt_sub_id.Text = "";
-            txt_score.Text = "";
-            
+
+            txt_id.ResetText();
+            txt_sub_id.ResetText();
+            txt_score.ResetText();
+
         }
 
-       
+
     }
 }
